Expose minimal gm_rdt_ training api
This starts making the training code available via a shared api and
with the train_rdt command line tool being built on that api.

As part of this change I ended up down a rat hole of signed/unsigned
type mismatch errors + found some inconsistent use of sized types. (In
part this was related to exposing state via gm_ui_properties which
support int types but but not specifically [u]int32_t)

This patch makes some broad stroke changes to avoid lots of sized types
where it seems reasonable to use an int for 99% of the misc integer
variables we have in our code where we aren't packing data tightly for
memory optimizations for for serializing (also not be fussy about using
unsigned integers for things like widths/heights considering C's foot
guns when mixing signed/unsigned arithmetic).

A minor change to the train_rdt command line interface was to drop the
'limit', 'skip' and 'shuffle' options considering the tool loads a
specified index where it's easy to control the size / shuffle of the
index instead.

+++ src/infer.h
@@ -47,19 +47,19 @@ typedef struct {
 
 template<typename FloatT>
 float* infer_labels(RDTree** forest,
-                    uint8_t n_trees,
+                    int n_trees,
                     FloatT* depth_image,
-                    uint32_t width,
-                    uint32_t height,
+                    int width,
+                    int height,
                     float* out_labels = NULL,
                     bool use_threads = false);
 
 template<typename FloatT>
 float* calc_pixel_weights(FloatT* depth_image,
                           float* pr_table,
-                          int32_t width,
-                          int32_t height,
-                          uint8_t n_labels,
+                          int width,
+                          int height,
+                          int n_labels,
                           JSON_Value* joint_map,
                           float* out_weights = NULL);
 
@@ -67,9 +67,9 @@ template<typename FloatT>
 InferredJoints* infer_joints_fast(FloatT* depth_image,
                                   float* pr_table,
                                   float* weights,
-                                  int32_t width,
-                                  int32_t height,
-                                  uint8_t n_labels,
+                                  int width,
+                                  int height,
+                                  int n_labels,
                                   JSON_Value* joint_map,
                                   float vfov,
                                   JIParam* params);
@@ -78,9 +78,9 @@ template<typename FloatT>
 InferredJoints* infer_joints(FloatT* depth_image,
                              float* pr_table,
                              float* weights,
-                             int32_t width,
-                             int32_t height,
-                             uint8_t n_labels,
+                             int width,
+                             int height,
+                             int n_labels,
                              JSON_Value* joint_map,
                              float vfov,
                              JIParam* params);
@@ -89,18 +89,18 @@ void free_joints(InferredJoints* joints);
 
 template<typename FloatT>
 float* reproject(FloatT* depth_image,
-                 int32_t width,
-                 int32_t height,
+                 int width,
+                 int height,
                  float vfov,
                  float threshold,
-                 uint32_t* n_points,
+                 int* n_points,
                  float* out_points = NULL);
 
 template<typename FloatT>
 FloatT* project(float* point_cloud,
-                uint32_t n_points,
-                int32_t width,
-                int32_t height,
+                int n_points,
+                int width,
+                int height,
                 float vfov,
                 float background = 0.f,
                 FloatT* out_depth = NULL);
